+++ py/desispec/test/test_qa.py
@@ -159,7 +159,7 @@ class TestQA(unittest.TestCase):
     def test_qa_exposure_load_write_data(self):
         #- Test loading data
         self._write_qaframes()
-        qaexp = QA_Exposure(self.expid, self.night, 'dark', specprod_dir=self.testDir)
+        qaexp = QA_Exposure(self.expid, self.night, 'science', specprod_dir=self.testDir)
         assert 'b0' in qaexp.data['frames']
         assert 'b1' in qaexp.data['frames']
         # Write